[FEATURE] Organize detecting SVs methods - move sa analysing methods
