+++ include/detect_breakends/bam_functions.hpp
@@ -2,6 +2,7 @@
 
 #include <seqan3/alphabet/cigar/cigar.hpp>
 #include <seqan3/utility/char_operations/predicate.hpp>
+#include <seqan3/io/alignment_file/output.hpp>
 #include <seqan3/io/exception.hpp>
 
 enum BamFlags
