+++ src/detect_breakends/junction_detection.cpp
@@ -7,111 +7,10 @@
 #include "detect_breakends/bam_functions.hpp"                       // for hasFlag* functions
 #include "modules/clustering/simple_clustering_method.hpp"          // for the simple clustering method
 #include "modules/sv_detecting_methods/analyze_cigar_method.hpp"    // for the split read method
-#include "structures/aligned_segment.hpp"                           // for struct AlignedSegment
+#include "modules/sv_detecting_methods/analyze_sa_tag_method.hpp"   // for the cigar string method
 #include "structures/cluster.hpp"                                   // for class Cluster
-#include "structures/junction.hpp"                                  // for class Junction
 
 using seqan3::operator""_tag;
-
-/*! \brief Splits a string by a given delimiter and stores substrings in a given container.
- *
- * \param str   string to split
- * \param cont  container for the splitted substrings
- * \param delim delimiter
- */
-template <class Container>
-void split_string(const std::string& str, Container& cont, char delim = ' ')
-{
-    std::stringstream ss(str);
-    std::string token;
-    while (std::getline(ss, token, delim)) {
-        cont.push_back(token);
-    }
-}
-
-/*! \brief Parse the SA tag from the SAM/BAM alignment of a chimeric/split-aligned read. Build
- *         [aligned_segments](\ref AlignedSegment), one for each alignment segment of the read.
- *
- * \param sa_string         "SA" tag string
- * \param aligned_segments  vector of [aligned_segments](\ref AlignedSegment).
- *
- * \details The SA tag describes the alignments of a chimeric read and is like a small SAM within a SAM file:
- *          "SA:Z:(rname,pos,strand,CIGAR,mapQ,NM;)+"
- *          Each element (in parentheses) represents one alignment segment of the chimeric alignment formatted as
- *          a colon-delimited list.
- *          We add all segments to our candidate list `aligned_segments` and examine them in the following function
- *          `analyze_aligned_segments()`.
- *          For more information about this tag, see the
- *          ([Map Optional Fields Specification](https://github.com/samtools/hts-specs/blob/master/SAMtags.pdf)).
- */
-void retrieve_aligned_segments(std::string sa_string, std::vector<AlignedSegment> & aligned_segments)
-{
-    std::vector<std::string> sa_tags{};
-    split_string(sa_string, sa_tags, ';');
-    for (std::string sa_tag : sa_tags)
-    {
-        std::vector<std::string> fields {};
-        split_string(sa_tag, fields, ',');
-        if (fields.size() == 6)
-        {
-            std::string ref_name = fields[0];
-            int32_t pos = std::stoi(fields[1]);
-            strand orientation;
-            if (fields[2] == "+")
-            {
-                orientation = strand::forward;
-            }
-            else if (fields[2] == "-")
-            {
-                orientation = strand::reverse;
-            }
-            else
-            {
-                continue;
-            }
-            std::string cigar_field = fields[3];
-            std::tuple<std::vector<seqan3::cigar>, int32_t, int32_t> parsed_cigar = parse_cigar(cigar_field);
-            std::vector<seqan3::cigar> cigar_vector = std::get<0>(parsed_cigar);
-            int32_t mapq = std::stoi(fields[4]);
-            aligned_segments.push_back(AlignedSegment{orientation, ref_name, pos, mapq, cigar_vector});
-        }
-    }
-}
-
-/*! \brief Build junctions out of aligned_segments.
- *
- * \param aligned_segments  vector of [aligned_segments](\ref AlignedSegment).
- * \param junctions         vector for storing junctions
- * \param read_name         QNAME field of the SAM/BAM file
- */
-void analyze_aligned_segments(const std::vector<AlignedSegment> & aligned_segments,
-                              std::vector<Junction> & junctions,
-                              const std::string & read_name)
-{
-    for(size_t i = 1; i<aligned_segments.size(); i++)
-    {
-        AlignedSegment current = aligned_segments[i-1];
-        AlignedSegment next = aligned_segments[i];
-        int32_t distance_on_read = next.get_query_start() - current.get_query_end();
-        // Neither gap nor overlap on read
-        if (distance_on_read >= -10 && distance_on_read <= 10)
-        {
-            Breakend mate1{current.ref_name,
-                            current.orientation == strand::forward ? current.get_reference_end()
-                                                                   : current.get_reference_start(),
-                            current.orientation,
-                            sequence_type::reference};
-            Breakend mate2{next.ref_name,
-                            next.orientation == strand::forward ? next.get_reference_start()
-                                                                : next.get_reference_end(),
-                            next.orientation,
-                            sequence_type::reference};
-            Junction new_junction{std::move(mate1), std::move(mate2), read_name};
-            seqan3::debug_stream << "BND: " << new_junction << "\n";
-            junctions.push_back(std::move(new_junction));
-        }
-    }
-}
 
 /*! \brief Detects junctions between distant genomic positions by analyzing an alignment file (sam/bam). The detected
  *         junctions are printed on stdout and insertion alleles are stored in a fasta file.
